ref(storage.go,storage_creds.go): move basic storage logic into convenience package
